God damn bug fixed
