+++ packages/ui/package.json
@@ -81,6 +81,7 @@
     "react-dom": "^18.2.0",
     "react-error-boundary": "^4.1.2",
     "react-helmet": "^6.1.0",
-    "react-player": "3.0.0-canary.3"
+    "react-player": "3.0.0-canary.3",
+    "screenfull": "^6.0.2"
   }
 }
