+++ packages/ui/src/universal/responsive/index.ts
@@ -3,7 +3,7 @@ import useMediaQuery from '@mui/material/useMediaQuery';
 
 const useIsMobile = () => {
   const theme = useTheme();
-  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+  const isMobile = useMediaQuery(theme.breakpoints.down('sm')); // sm is 600px by default
 
   return isMobile;
 };
