+++ packages/ui/src/watch/video-detail-page/containers/skeleton.tsx
@@ -1,0 +1,5 @@
+const LoadingSkeleton = () => {
+  return null;
+};
+
+export { LoadingSkeleton };
